requirements.txt
streamlit==1.24.1
matplotlib==3.7.1
numpy==1.25.0
google-generativeai==0.1.0
Sympy


